refactor product table code
